Add changes

+++ src/database/DataTypes/TransactionBatch.js
@@ -167,7 +167,7 @@ TransactionBatch.schema = {
     sellPrice: 'double?',
     donor: { type: 'Name', optional: true },
     sortIndex: { type: 'int', optional: true },
-    total: { type: 'float', optional: true },
+    total: { type: 'double', optional: true },
     type: { type: 'string', optional: true },
     linkedTransaction: { type: 'Transaction', optional: true },
   },
